Fix up warning-as-error on Windows

+++ src/Search/Swallow/SdkModel/SwallowSearchMenuModule.cpp
@@ -68,7 +68,7 @@ namespace ExampleApp
                                                         Eegeo_NEW(View::SwallowCategoryMenuOption)(Search::Swallow::SearchConstants::OFFICE_CATEGORY_NAME,
                                                                                                    false,
                                                                                                    menuViewModel,
-                                                                                                   Eegeo::Space::EarthConstants::Radius*2.0,
+                                                                                                   static_cast<float>(Eegeo::Space::EarthConstants::Radius*2.0),
                                                                                                    messageBus));
                 }
